refactor: update file names to be more deno-like
