+++ source/models/directory_core.ts
